Fixed error in reading function
Because of type casting it would think there was
errors if the byte sent had a value of 255 because
for a 'char' it is -1 but for an 'int' it is 255.

+++ Seriel project/SLIP/slip.h
@@ -39,6 +39,6 @@ void sendPacket(char* data, size_t n);
 *	\pre A connection is open. Data can contain atleast DATA_MAX.
 *	\post 'data' now contains the data from the package and the number of bytes received is returned.(0 if error stream or timeout)
 */
-size_t receivePackage(char data[]);
+int receivePackage(char data[]);
 
 #endif
